+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Breaking changes are distinguised by this symbol: 🔧
 
 - Removed outdated Agents Query Category 🔧
 - Added option to specify Ascending or Descending sort
+- Updated Table Columns selection component in the query editor
 
 ## [1.2.3] - 2021-12-10